Add low level _log package for internal use
Internally, the "log" package wants to use the "config" package,
which in turn, wants to use the "log" package, creating a
cyclical dependency. To resolve this, we build a logger in a lower
level "_log" package, with a default config, which will be used by
other packages. When the "log" package is initialized, it will
overwrite the configuration in the "_log" package, and proxy through
the "_log" package functions to its own interface.
The trick however, is to ideally only begin logging when the "log"
package has been set up, so that all emitted logs follow the users
configuration. Therefore, before the "log" package is set up, the
lower level "_log" package, will stash all logs without emitting them,
until the "log" package tells it that it has been configured,
or until the program terminates. If the program terminates, we emit
all logs in the stash, regardless of whether the "log" package
has been configured or not.